+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - added podcasts (disabled by default).
+- praise the linux gods, MPRIS Player support is here.
 - we're aware of Lastfm API Changes, we removed all of their APIs from Dawayer as a response, sorry for the inconvenience, we will have a solution soon.
 - support for opening a track file while Dawayer is already running.
 - fixed an error that appears when scanning a non-existing directories.